v5-core/v4-sync-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
kafka-python==2.0.2
opensearch-py==2.4.2
psycopg2-binary==2.9.9
prometheus-client==0.19.0
python-json-logger==2.0.7
